[tests] add support for tests using Bendy-Bus

+++ tests/runtest.sh
@@ -10,6 +10,18 @@ then
     exit 1
 fi
 
-eval `dbus-launch --sh-syntax`
-trap "kill ${DBUS_SESSION_BUS_PID}" EXIT
-./${TEST_NAME}.bin "${@}"
+if [[ -e ${TEST_NAME}.machine ]]
+then
+    BB_ARGS="${BB_ARGS} --run-iters=1 --pass-through-environment"
+
+    # There must be bug in bendy-bus: with the switch '--pass-through-environment'
+    # DBUS_SESSION_BUS_ADDRESS instead of being replaced appears **twice** in process environment.
+    unset DBUS_SESSION_BUS_ADDRESS
+
+    shift
+    bendy-bus ${BB_ARGS} ${TEST_NAME}.machine interfaces.xml -- $(pwd)/${TEST_NAME}.bin "${@}"
+else
+    eval `dbus-launch --sh-syntax`
+    trap "kill ${DBUS_SESSION_BUS_PID}" EXIT
+    ./${TEST_NAME}.bin "${@}"
+fi
